Moved formulize to src folder
