+++ src/formulize.ts
@@ -7,11 +7,11 @@
 import {
   FormulaStore,
   formulaStoreManager,
-} from "../store/FormulaStoreManager";
-import { computationStore } from "../store/computation";
-import { IEnvironment } from "../types/environment";
-import { IVariable } from "../types/variable";
-import { getVariable } from "../util/computation-helpers";
+} from "./store/FormulaStoreManager";
+import { computationStore } from "./store/computation";
+import { IEnvironment } from "./types/environment";
+import { IVariable } from "./types/variable";
+import { getVariable } from "./util/computation-helpers";
 
 export interface FormulizeConfig extends IEnvironment {}
